refactor: apply requested changes
