+++ packages/ui/src/components/TeamCard/TeamCard.tsx
@@ -6,6 +6,9 @@ import { Box } from '../Box';
 import { Link } from '../Link';
 import { Typography } from '../Typography';
 
+export const NOT_ACTIVE_TEXT =
+  'Jeszcze nie działasz w żadnym zespole. Zostaniesz do niego przydzielony automatycznie po napisaniu pierwszego testu i zakwalifikowaniu się na kurs.';
+
 export type TeamMember = {
   id: number;
   name: string;
@@ -19,22 +22,23 @@ export interface TeamCardProps extends BoxProps {
   active: boolean;
 }
 
-export const notActiveText =
-  'Jeszcze nie działasz w żdanym zespole. Zostaniesz do niego przydzielony automatycznie po napisaniu pierwszego testu i zakwalifikowaniu się na kurs.';
+type AvatarProps = {
+  teamMembers: TeamMember[];
+};
+
+const Avatars = ({ teamMembers }: AvatarProps) => {
+  return (
+    <AvatarGroup spacing="24px" marginTop="4px">
+      {teamMembers.map(({ id, name, image, profileUrl }) => (
+        <Link href={profileUrl} key={id}>
+          <Avatar src={image} name={name} size="sm" />
+        </Link>
+      ))}
+    </AvatarGroup>
+  );
+};
 
 export const TeamCard = ({ members, mentors, active, ...props }: TeamCardProps) => {
-  const renderAvatars = (teamMembers: TeamMember[]) => {
-    return (
-      <AvatarGroup spacing="24px">
-        {teamMembers.map(({ id, name, image, profileUrl }) => (
-          <Link href={profileUrl} key={id}>
-            <Avatar src={image} name={name} size="sm" />
-          </Link>
-        ))}
-      </AvatarGroup>
-    );
-  };
-
   return (
     <Box width="100%" boxShadow="base" borderRadius={8} padding="24px" {...props}>
       <Typography size="2xl" fontWeight={800} lineHeight="28px">
@@ -46,15 +50,15 @@ export const TeamCard = ({ members, mentors, active, ...props }: TeamCardProps) 
             <Typography fontWeight="medium" size="md">
               Członkowie
             </Typography>
-            {renderAvatars(members)}
+            <Avatars teamMembers={members} />
             <Typography fontWeight="medium" size="md" marginTop="16px">
               Mentor
             </Typography>
-            {renderAvatars(mentors)}
+            <Avatars teamMembers={mentors} />
           </>
         ) : (
           <Typography fontWeight="400" lineHeight="24px">
-            {notActiveText}
+            {NOT_ACTIVE_TEXT}
           </Typography>
         )}
       </Box>
