+++ packages/ui/src/theme/overwrites/components/index.ts
@@ -1,4 +1,5 @@
 import { Input } from './input';
+import { Link } from './link';
 import { NumberInput } from './numberInput';
 import { Radio } from './radio';
 import { Textarea } from './textarea';
@@ -8,4 +9,5 @@ export const components = {
   NumberInput,
   Textarea,
   Input,
+  Link,
 };
