+++ packages/ui/src/theme/overwrites/components/link.ts
@@ -1,0 +1,7 @@
+export const Link = {
+  baseStyle: {
+    _hover: {
+      textDecoration: 'none',
+    },
+  },
+};
